Complete Month 1: Sanctum auth with fresh migrations, AuthController, and mobile app setup

+++ backend/app/database/migrations/2026_06_27_000004_create_budgets_table.php
@@ -1,37 +1,29 @@
 <?php
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('budgets', function (Blueprint $table) {
             $table->id();
-
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->foreignId('category_id')->constrained()->cascadeOnDelete();
-
             $table->decimal('limit_amount', 10, 2);
-            $table->tinyInteger('month');
-            $table->integer('year');
-
+            $table->integer('month'); // 1-12
+            $table->integer('year'); // e.g., 2026
             $table->timestamps();
-
+            
+            // User can only set one budget per category per month
             $table->unique(['user_id', 'category_id', 'month', 'year']);
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('budgets');
     }
 };
-
